% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caaqs_stn_history.R
\name{add_TFEE}
\alias{add_TFEE}
\title{check the data for TFEE, and add TFEE flag}
\usage{
add_TFEE(
  df,
  station_column = NULL,
  date_column = NULL,
  param_column = NULL,
  is_time_ending = TRUE
)
}
\arguments{
\item{df}{is the dataframe containing hourly or daily data}

\item{station_column}{is a string defining the station name.
By default it will be the column containg "station" or "site" in the name}

\item{date_column}{is the column containing the date.
Note that if this is date/time column, is_time_ending should be specified}

\item{param_column}{is the column defining the pollutant or parameter}

\item{is_time_ending}{is either TRUE/FALSE. default is TRUE.
time_ending means that the start of the day is AFTER and NOT including midnight.}
}
\description{
This function adds TFEE flag to hourly or daily data
}
\examples{
df <- importBC_data('pm25',years = 2020)
df_tfee <- add_TFEE(df)
df_tfee <- add_TFEE(df, station_column = 'STATION_NAME',date_column = 'DATE_PST',is_time_ending = TRUE)
}
