% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairecms.R
\name{CLEAN_FTP}
\alias{CLEAN_FTP}
\title{Clean up FTP Files}
\usage{
CLEAN_FTP(URL, days_to_keep = 365)
}
\arguments{
\item{URL}{is vector of FTP URL, includes ftp://}

\item{days_to_keep}{the age of data that will be retained, any older are deleted}
}
\description{
This function removed old files in the FTP URL
}
