% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBC_stations.R
\name{clean_stationnames}
\alias{clean_stationnames}
\title{This function clean up station names
Removes extra spaces, and non-alphanumeric characters}
\usage{
clean_stationnames(df, cols = c("STATION_NAME", "STATION_NAME_FULL"))
}
\description{
This function clean up station names
Removes extra spaces, and non-alphanumeric characters
}
