% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{direction}
\alias{direction}
\title{Get the compass needle value of an azimuth measurement}
\usage{
direction(angle)
}
\arguments{
\item{angle}{is a vector of numbers, corresponds to azimuth readings}
}
\description{
The result is OVER or UNDER if it is not inside 0-360 deg value
Note: must be efficient if this function revised to use Mod and div instead
}
