% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_process.R
\name{download_files}
\alias{download_files}
\title{Download a list of FTP files}
\usage{
download_files(url_list, save_dir = NULL)
}
\arguments{
\item{url_list}{is a vector containing list of URLs to download}

\item{save_dir}{is the directory where files are saved, NULL means temporary filenames used}
}
\description{
uses parallel CPU processing to perform simultaneous downloads
}
