% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventingBC_bulletin.R
\name{extract_file_date}
\alias{extract_file_date}
\title{Extract the date based on the filename specified in FTP}
\usage{
extract_file_date(filenames)
}
\description{
Extract the date based on the filename specified in FTP
}
