% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbc_data.R
\name{extractDateTime}
\alias{extractDateTime}
\title{Extracts the date time string into a format YYYY-MM-DD hh:mm}
\usage{
extractDateTime(dateTimeString)
}
\arguments{
\item{dateTimeString}{is a vector containing datetime string}
}
\description{
Extracts the date time string into a format YYYY-MM-DD hh:mm
}
