% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{FILL_Recent_Data}
\alias{FILL_Recent_Data}
\title{Forward fills data with previous value}
\usage{
FILL_Recent_Data(data_all, threshold = 2)
}
\arguments{
\item{data_all}{is the dataframe containing all data in original wide format}

\item{threshold}{is the number of hours of maximum delay}
}
\description{
This addresses the issue where a station has multiple dataloggers
and some dataloggers have not updated recent data
Works only for 1-station entry, no STATION_NAME_FULL
}
