% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{FIX_tooltip}
\alias{FIX_tooltip}
\title{Fix tooltip in rmarkdown kable
This fixes tooltip issue by creating new lines
by inserting spaces}
\usage{
FIX_tooltip(txt, sep_string = "<br>")
}
\arguments{
\item{txt}{is a string containing the tooltip
lines are separated by sep_string}

\item{sep_string}{is the string separator for the lines
default is <br>}
}
\description{
Fix tooltip in rmarkdown kable
This fixes tooltip issue by creating new lines
by inserting spaces
}
