% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caaqs_result.R
\name{get_annual_metrics}
\alias{get_annual_metrics}
\title{Get Annual metrics}
\usage{
get_annual_metrics(parameter, years)
}
\arguments{
\item{parameter}{is the pollutant. Only one pollutant at a time}

\item{years}{are the years where the annual value is calculated}
}
\description{
Same as importBC_data_avg, added for simplicity and consistent
}
\examples{
get_annual_metrics("pm25", years = 2017:2020)

}
