% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emissioninventory.R
\name{get_apei}
\alias{get_apei}
\title{Retrieve the air pollutant emission inventory from ECCC datamart}
\usage{
get_apei(categorytype = "Source", pollutant = NULL, df = NULL, URL = NULL)
}
\arguments{
\item{categorytype}{is either source, sector, or subsector. Default is source}

\item{pollutant}{is the pollutant to retrieve. If NULL, it will retrieve all pollutants}

\item{df}{is the dataframe containg the data from source. added to expedite for the shiny app version.}

\item{URL}{the ECCC URL for the NPRI. Can be downloaded file in local directory or in ECCC data mart.
At writing, it datamart location is at:
'https://data-donnees.ec.gc.ca/data/substances/monitor/canada-s-air-pollutant-emissions-inventory/EN_APEI-Can-Prov_Terr.csv'}
}
\description{
Files are retrieved from ECCC APRI site
}
