% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caaqs_result.R
\name{get_caaqs}
\alias{get_caaqs}
\title{Define the caaqs values}
\usage{
get_caaqs(years = NULL)
}
\arguments{
\item{years}{are the years to be included}
}
\description{
Define the caaqs values
}
\examples{
get_caaqs(years = 2010:2030)

}
