% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caaqs_result.R
\name{get_caaqs_metrics}
\alias{get_caaqs_metrics}
\title{Calculates the CAAQS values for a particular year and parameter}
\usage{
get_caaqs_metrics(parameter, years)
}
\arguments{
\item{parameter}{is the pollutant that will be processed, one pollutant at a time.}

\item{years}{are the years that will be included in the calculation of CAAQS values}
}
\description{
Calculates the CAAQS values for a particular year and parameter
}
\examples{
get_caaqs_metrics("pm25", years = 2017:2020)

}
