% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_captures.R
\name{get_captures0}
\alias{get_captures0}
\title{backend function}
\usage{
get_captures0(
  param,
  years = NULL,
  merge_Stations = FALSE,
  stop_at_present = TRUE
)
}
\description{
This function makes all the needed calculations to generate the captures data
}
