% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairecccdata.R
\name{GET_CGNDB}
\alias{GET_CGNDB}
\title{INCOMPLETE: Get the CGNDB of a specified city}
\usage{
GET_CGNDB(list.cities)
}
\arguments{
\item{list.cities}{is a string vector of cities}
}
\description{
This function retrieves CGNDB details
}
