% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_completeness.R
\name{get_data_completeness}
\alias{get_data_completeness}
\title{Calculate the data capture for the station and parameter}
\usage{
get_data_completeness(
  parameter,
  years = NULL,
  groupby = NULL,
  merge_Stations = FALSE,
  stop_at_present = TRUE
)
}
\arguments{
\item{parameter}{is the pollutant(param). Use list_parameters() to list availble parameters.
For one pollutant only
param can also be a dataframe retrieved from an importBC_data() function.}

\item{years}{is the year or years. You can use vectors to query multiple years. If NULL, it will be the current year}

\item{groupby}{is vector listing the columns that the data is grouped by, if null
it groups station and instrument for PM
groups by station for non-PM}

\item{merge_Stations}{default is FALSE. if TRUE, it will combine some stations/instruments
based on the merging requirement applies in CAAQS}

\item{stop_at_present}{default is TRUE if TRUE, it will stop the date at the present date}
}
\description{
NOTE: This will output the valid hours, valid days, valid days in quarter, valid Q2+Q3, valid
}
\examples{
get_data_completeness(c('pm25','no2'), years = 2015:2020,merge_Stations = TRUE)

}
