% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairecccdata.R
\name{GET_ECCC_AQHI}
\alias{GET_ECCC_AQHI}
\title{Get AQHI from ECCC}
\usage{
GET_ECCC_AQHI(type = "REALTIME")
}
\arguments{
\item{type}{is either REALTIME or FORECAST, or FORECASTSUMMARY}
}
\description{
Retrieves hourly AQHI from the ECCC datamart website
}
