% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairecccdata.R
\name{GET_ECCC_FORECAST}
\alias{GET_ECCC_FORECAST}
\title{GET ECCC FORECAST function}
\usage{
GET_ECCC_FORECAST(parameter = NULL)
}
\arguments{
\item{parameter}{is a vector string of AQHI, PM25, NO2, O3, PM10. It will include all forecase if NULL}
}
\description{
This function retrieves latest forecast from ECCC models
}
