% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caaqs_stn_history.R
\name{get_excel_table}
\alias{get_excel_table}
\title{Retrieve excel sheet tables from BC ENV Open data Portal}
\usage{
get_excel_table(
  ExcelURL,
  sheet = NULL,
  header_row = NULL,
  data_row = NULL,
  silent = FALSE
)
}
\arguments{
\item{ExcelURL}{is the complete URL of the excel file.}

\item{sheet}{defines the worksheet names in the excel file. can be a vector listing sheets
for vector entry (multiple sheets), it will combine result into a single table
if NULL and excel file only has one worksheet, it will process that worksheet.
If NULL and there are multiple worksheets, function will return with list of worksheets}

\item{header_row}{specifies which row is the header.
If NULL, the function will scan the most likely header row}

\item{data_row}{specifies which row is the first data start. Typically header_row +1
If NULL, the function assumes the row below the header_row}
}
\description{
This function helps access excel sheets table that are located in FTP Open Data Portal
}
\examples{
get_excel_table('ftp://ftp.env.gov.bc.ca/pub/outgoing/AIR/CAAQS/BC_CAAQS_station_history.xlsx')
get_excel_table('ftp://ftp.env.gov.bc.ca/pub/outgoing/AIR/CAAQS/BC_CAAQS_station_history.xlsx',sheet = c('TFEE PM2.5','TFEE O3'))
}
