% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairecccdata.R
\name{GET_FILE_CSV_deprecated}
\alias{GET_FILE_CSV_deprecated}
\title{GET_FILE_CSV}
\usage{
GET_FILE_CSV_deprecated(list.files, ftp.path, path.local = NULL, clean = TRUE)
}
\arguments{
\item{list.files}{is a vector string containing filesnames with extension}

\item{ftp.path}{is the ftp path where data will be saved}

\item{path.local}{is the local directory where the file will be saved. Default is NULL where temp folder is in the workpath}
}
\description{
This function retrieves csv file from ftp, save local copy. It only retrieves updated files
}
