% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_FTP_DETAILS}
\alias{GET_FTP_DETAILS}
\title{GET FTP DETAILS}
\usage{
GET_FTP_DETAILS(path.ftp)
}
\arguments{
\item{path.ftp}{the ftp path}
}
\description{
grabs the file and directory details of an ftp
}
