% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_FTP_DETAILS_}
\alias{GET_FTP_DETAILS_}
\title{primary basic function for GET_FTPDetails
added here to accommodate multiple list}
\usage{
GET_FTP_DETAILS_(path.ftp)
}
\description{
primary basic function for GET_FTPDetails
added here to accommodate multiple list
}
