% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_process.R
\name{get_ftp_parallel}
\alias{get_ftp_parallel}
\title{Download a list of FTP files}
\usage{
get_ftp_parallel(url_list)
}
\arguments{
\item{url_list}{is a vector containing list of URLs to download}
}
\description{
uses parallel CPU processing to perform simultaneous downloads
}
\details{
NOTE: This is found significantly slower than using GET_FTP_DETAILS() on multiple FTP files
}
