% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_LOCAL_Time}
\alias{GET_LOCAL_Time}
\title{Convert to Local Pacific Timezone}
\usage{
GET_LOCAL_Time(DATE_PST, format = "\%Y-\%m-\%d \%H:\%M")
}
\arguments{
\item{DATE_PST}{is vector of PST date-time in ymd_hm format}

\item{format}{is the format of the output}
}
\description{
Convert to Local Pacific Timezone
}
