% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caaqs_result.R
\name{get_management}
\alias{get_management}
\title{Define the management levels of given it's parameter, metric, value, and year}
\usage{
get_management(df)
}
\arguments{
\item{df}{is the dataframe. Must have the following columns: parameter, year, metric,metric_value,tfee}
}
\description{
NOTE: updated until 2025-12-03
If CAAQS or management levels are updated, please update this and the get_caaqs() function
}
