% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairstatfunctions.R
\name{GET_NEGATIVE_FLAT}
\alias{GET_NEGATIVE_FLAT}
\title{Get negative and flat data reports}
\usage{
GET_NEGATIVE_FLAT(
  parameter = NULL,
  YEAR = NULL,
  file.result = "negative_flats.csv",
  exclude_MVRD = TRUE
)
}
\arguments{
\item{parameter}{default is NULL to include all relevant parameters
or vector listing all parameters}

\item{file.result}{is the location where resulting file is to be saved}

\item{exclude_MVRD}{means MVRD data is excluded}

\item{year}{is the year to check for negative and flat,
if NULL,}
}
\description{
Get negative and flat data reports
}
