% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_emissioninventory.R
\name{get_npri}
\alias{get_npri}
\title{Retrieve NPRI Emissions and Sources from ECCC Datamart}
\usage{
get_npri(URL_emissions = NULL, URL_sources = NULL)
}
\arguments{
\item{URL_emissions}{is the ECCC datamart download link for the NPRI release csv file}

\item{URL_sources}{is the ECCC datamart download link for the NPRI Emission sources}
}
\description{
Script here generates an output file based on the NPRI data----
NPRI Data is owned and maintained by Environment and Climate Change Canada
If the URL is incorrect, please refer to:
https://open.canada.ca/data/en/dataset/40e01423-7728-429c-ac9d-2954385ccdfb
}
