% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairstatfunctions.R
\name{GET_nTH_HIGHEST}
\alias{GET_nTH_HIGHEST}
\title{Retrieve the nth highest number}
\usage{
GET_nTH_HIGHEST(data.input, column.name = "ROUNDED_VALUE", data.order = 4)
}
\arguments{
\item{data.input}{is the dataframe input}

\item{column.name}{is the name of the column where the nth highest value is retrieved}

\item{data.order}{is the order of data}
}
\description{
Retrieve the nth highest number
}
