% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_PARAMETER_DATA}
\alias{GET_PARAMETER_DATA}
\title{GET PARAMETER DATA FUNCTION}
\usage{
GET_PARAMETER_DATA(data.parameter, year.start, year.end = NULL)
}
\arguments{
\item{data.parameter}{the air quality parameter. can be PM25, PM10, SO2, TRS, H2S, etc.}

\item{year.start}{the beginning of query, year only}

\item{year.end}{the end of query, year only. the same year if unspecified
GET_VALID_DATA_PARAMETER()}
}
\description{
This function retrieves parameter data from the FTP open data portal
Data includes verified and unverified data
}
\details{
Updated: 2020-02-13 with updates for reading the data, parsing fix
}
