% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairstatfunctions.R
\name{GET_PIVOT_TABLE}
\alias{GET_PIVOT_TABLE}
\title{Created before pivot_wider and pivot_longer became available}
\usage{
GET_PIVOT_TABLE(
  data.input,
  columnname.category,
  columnname.value,
  columnname.levels = NULL
)
}
\description{
Created before pivot_wider and pivot_longer became available
}
