% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_RECENT_HOURLY_DATA}
\alias{GET_RECENT_HOURLY_DATA}
\title{GET RECENT HOURLY DATA by PARAMETER/STATION FUNCTION}
\usage{
GET_RECENT_HOURLY_DATA(data.request)
}
\arguments{
\item{data.request}{vector of either parameters or station names for the data that will be retrieved.
Function automatically identify if you entered station name or parameter}
}
\description{
This function retrieves the hourly  data for a specific parameter. Contains only the recent 1-2 months
}
