% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_RECENT_MINUTE_DATA}
\alias{GET_RECENT_MINUTE_DATA}
\title{GET MINUTE DATA by PARAMETER/STATION FUNCTION}
\usage{
GET_RECENT_MINUTE_DATA(data.request)
}
\arguments{
\item{data.request}{parameter or station name for the data that will be retrieved.
Function automatically identify if you entered station name or parameter}
}
\description{
This function retrieves the minute data for a specific parameter. Contains only the recent 1-2 months
}
