% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_RECENT_STATION_DATA}
\alias{GET_RECENT_STATION_DATA}
\title{GET RECENT STATION DATA FUNCTION}
\usage{
GET_RECENT_STATION_DATA(STATION = "ALL", timebase = 60)
}
\arguments{
\item{STATION}{vector string of the air quality monitoring station, note this excludes}

\item{timebase}{default to 60, is either 1 or 60. If 1, function is same as GET_MINUTE_DATA_FTP
the underscore naming system, e.g., Prince George Plaza 400 NOT Prince George Plaza 400 Met_60
GET_RECENT_DATA_STATION_FTP()}
}
\description{
This function retrieves the recent 1 month  of data from the ftp feed
}
