% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_ROLLING_MEAN}
\alias{GET_ROLLING_MEAN}
\title{Get Rolling Mean function}
\usage{
GET_ROLLING_MEAN(
  data.input,
  data.input.previous = NULL,
  data.averaginghour = 8,
  column.date = "DATE_PST",
  column.value = "ROUNDED_VALUE",
  column.static = c("STATION_NAME", "STATION_NAME_FULL", "SERIAL", "EMS_ID",
    "INSTRUMENT"),
  threshold = 0.75,
  precision = 1
)
}
\arguments{
\item{data.input}{is dataframe containing data}

\item{data.input.previous}{is dataframe of earlier data, added to have complete rolling mean}

\item{data.averaginghour}{is the rolling mean time}

\item{column.date}{is string defining the date time column}

\item{column.value}{is vector that will be averaged}

\item{column.static}{is vector sting containing all columns where rolling average is calculated}

\item{threshold}{is the data capture requirement in fraction. If the valid sample
for average is less, it will not have a rolling mean value}

\item{precision}{is the number of decimal places for the result}
}
\description{
This function creates a rolling mean based on specified hours
NOTE: use GET_RUNNING_AVG() for tidy version
}
