% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_RUNNING_AVG}
\alias{GET_RUNNING_AVG}
\title{Calculate the running average}
\usage{
GET_RUNNING_AVG(
  airdata,
  parameter = NULL,
  date_column = "DATE_PST",
  parameter_column = "name",
  value_column = "value",
  averaging_time = 24,
  grouping_column = "STATION_NAME",
  threshold_percent = 75
)
}
\arguments{
\item{airdata}{is the dataframe containing the air quality
data. Requires columns listing the parameter and the value}

\item{parameter}{is vector contaling parameters listed in the parameter column
where the running average is calculated
if NULL, it will include each parameters in averaging}

\item{date_column}{is the column contains the date-time is ymd_hm format}

\item{parameter_column}{is the column listing all the paremetsr}

\item{value_column}{is the column containing the data value}

\item{averaging_time}{is the time to perform running average}

\item{grouping_column}{is the column where result will be grouped by}

\item{threshold_percent}{is the percent of data that are valid, otherwise
the averaging result is NA}
}
\description{
Gets the running average of specific parameters
requires a pivot_longer format
}
