% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_SAS_EXCEL_FILE}
\alias{GET_SAS_EXCEL_FILE}
\title{Retrieves the excel files Containing the Statistical Summary
This fixes the header where in the header is made of two rows}
\usage{
GET_SAS_EXCEL_FILE(filename)
}
\arguments{
\item{filename}{is the full path of the file}
}
\description{
Retrieves the excel files Containing the Statistical Summary
This fixes the header where in the header is made of two rows
}
