% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_STATION_DETAILS_FTP}
\alias{GET_STATION_DETAILS_FTP}
\title{GET STATION DETAILS FUNCTION}
\usage{
GET_STATION_DETAILS_FTP(data.year = as.character(format(Sys.Date(),
  "\%Y")))
}
\arguments{
\item{data.year}{the year where station details are retrieved from. Defaults to current year}
}
\description{
This function retrieves latest station details or deteails during specific year from the ftp feed
}
