% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairstatfunctions.R
\name{GET_STATISTICS_PARAMETER}
\alias{GET_STATISTICS_PARAMETER}
\title{Gets the statistical details of a parameter}
\usage{
GET_STATISTICS_PARAMETER(data.year, parameter,
  instrument.ignore = !(tolower(parameter) \%in\% c("pm25", "pm10")),
  data.source = NULL)
}
\arguments{
\item{data.year}{numeric year. includes unverified and validated year}

\item{parameter}{string or vector containing the parameter that will be evaluated.}

\item{instrument.ignore}{if specified, ignores the instrument technology and merges dataset during statistical calculations. Default is FALSE for particulate (PM25,PM10) instruments}
}
\description{
This function renames or deletes the column of a dataframe based on its name
}
\keyword{GET_STATISTICS_PARAMETER()}
\keyword{annual}
\keyword{data}
\keyword{data,}
\keyword{statistics,}
\keyword{unverified}
\keyword{valid}
