% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats_}
\alias{get_stats_}
\title{Calculate the statistics for the pollutants based on CAAQS}
\usage{
get_stats_(param, years = NULL, add_TFEE = FALSE, merge_Stations = FALSE)
}
\arguments{
\item{param}{is the parameter or vector of parameters.}

\item{add_TFEE}{default FALSE. If TRUE, it will also calculate on data without TFEE}

\item{merge_Stations}{default FALSE. If TRUE, it will combine stations as practiced in air zone reporting}

\item{datetime}{is a string defining the datetime field. This field shoule be in time-ending format.}
}
\description{
Note that these are based on the metrics defined by the CCME
Guidance Document of Achievement Determination
}
