% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caaqs_stn_history.R
\name{get_tfee}
\alias{get_tfee}
\title{Retrieve TFEE data}
\usage{
get_tfee(ExcelURL = NULL)
}
\arguments{
\item{ExcelURL}{is the complete uRL of the excel file containing TFEE data
leave NULL to use known location of fTP at 'ftp://ftp.env.gov.bc.ca/pub/outgoing/AIR/CAAQS/BC_CAAQS_station_history.xlsx'}
}
\description{
This function retrieve TFEE data from the FTP
}
\examples{
df_tfee <- get_tfee()
}
