% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_URL_FOLDERS}
\alias{GET_URL_FOLDERS}
\title{Get Folders in URL}
\usage{
GET_URL_FOLDERS(source.url = "http://dd.weather.gc.ca/bulletins/alphanumeric/")
}
\arguments{
\item{source.url}{is the URL containing the data folders, default is ECCC datamart}
}
\description{
This function retrieves the list of folders that are in the specified URL
}
