% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairecccdata.R
\name{GET_URL_FOLDERS_deprecated}
\alias{GET_URL_FOLDERS_deprecated}
\title{Get Folders in URL}
\usage{
GET_URL_FOLDERS_deprecated(
  source.url = "http://dd.weatheroffice.ec.gc.ca/bulletins/alphanumeric/"
)
}
\arguments{
\item{source.url}{is the URL containing the data folders, default is ECCC datamart}
}
\description{
This function retrieves the list of folders that are in the specified URL
}
