% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairstatfunctions.R
\name{GET_VALID_COUNT}
\alias{GET_VALID_COUNT}
\title{Counts the number of valid hours or days}
\usage{
GET_VALID_COUNT(data.hourly, day.threshold = 0.75, precision = 1,
  Q2Q3only = FALSE)
}
\arguments{
\item{Q2Q3}{only
GET_VALID_COUNT()}
}
\description{
This function counts the number of valid hours
}
