% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{GET_VALID_DATA_PARAMETER}
\alias{GET_VALID_DATA_PARAMETER}
\title{GET VALID PARAMETER DATA FUNCTION}
\usage{
GET_VALID_DATA_PARAMETER(data.parameter, year.start, year.end = NULL)
}
\arguments{
\item{data.parameter}{the air quality parameter. can be PM25, PM10, SO2, TRS, H2S, etc.}

\item{year.start}{the beginning of query, year only}

\item{year.end}{the end of query, year only. Latest valid year if unspecified
GET_VALID_DATA_PARAMETER()}
}
\description{
This function retrieves parameter data from the open data portal
}
