% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventingBC_bulletin.R
\name{get_venting}
\alias{get_venting}
\title{get ventilation index
Update: 2026-04-29}
\usage{
get_venting(dates_included)
}
\arguments{
\item{dates_included}{is the sequence of dates to include}
}
\description{
retrieve data from FTP
ftp://ftp.env.gov.bc.ca/pub/outgoing/Air/VentingBulletins/\if{html}{\out{<yyyy>}}
}
