% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventingBC_bulletin.R
\name{GET_VENTING_ECCC0}
\alias{GET_VENTING_ECCC0}
\title{This is now a backend function
use GET_VENTING_ECCC()}
\usage{
GET_VENTING_ECCC0(date.start = NULL)
}
\description{
This is now a backend function
use GET_VENTING_ECCC()
}
