% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairventingindex.R
\name{GET_VENTING_OBSCR}
\alias{GET_VENTING_OBSCR}
\title{Creates kml files or shapefiles}
\usage{
GET_VENTING_OBSCR(path.output = NULL)
}
\arguments{
\item{path.output}{where kml files will be saved. If null, function returns shape file}
}
\description{
This function creates kml files using the new OBSCR sensitivity zones and venting regions
}
