% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventingBC_bulletin.R
\name{get_venting_summary}
\alias{get_venting_summary}
\title{get_venting_monthly_summary}
\usage{
get_venting_summary(date_from, date_to)
}
\arguments{
\item{date_from}{is a the starting dates, as string in yyyy-mm or yyyy-mm-dd}

\item{date_to}{is the end dates, as string in yyyy-mm or yyyy-mm-dd}
}
\description{
retrieves the monthly summary of today's ventilation index forecast
}
