% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbc_data.R
\name{importBC_data}
\alias{importBC_data}
\title{Import Hourly BC Data from station or parameter}
\usage{
importBC_data(
  parameter_or_station,
  years = NULL,
  flag_TFEE = TRUE,
  merge_Stations = TRUE,
  clean_names = TRUE,
  use_openairformat = TRUE
)
}
\arguments{
\item{parameter_or_station}{vector list of air quality parameters or station (automatic).
use the function listBC_stations() to get a detailed list of BC Air Quality Monitoring stations
if there is no exact station match, it will use partial match
so that user can just enter 'Prince George' and it will search for all stations
with Prince  George in the station name (Prince George Plaza 400, Prince George Glenview)
List of parameters can be obtained using list_parameters() command
List of stations can be retrieved using listBC_stations() command
Mutliple stations can be specified}

\item{years}{the years that will be retrieved. For sequence, use 2009:2015. For non
sequential years, use c(2010,2015,2018)
If not declared, the current year will be used}

\item{flag_TFEE}{default is FALSE. If TRUE, it will add TFEE flags on days TFEE were verified}

\item{merge_Stations}{default is FALSE. it will combine data from stations and alternative stations}

\item{clean_names}{makes the output columns in lower case letters as acceptable with tidyverse
This function retrieves thsoe details from the CAAQS station history excel file
ftp://ftp.env.gov.bc.ca/pub/outgoing/AIR/CAAQS/}

\item{use_openairformat}{is boolean,if TRUE, output is compatible with openair. Applies only to station queries}
}
\description{
This function retrieves station or parameter hourly data from the BC open data portal
Data includes verified and unverified data depending on whether Level 2 has been completed for that date
}
\examples{
importBC_data('Prince George Plaza 400')
importBC_data('pm25',2015:2016,use_openairformat = FALSE)
importBC_data(c('Prince George','Kamloops'),c(2010,2015))

}
