% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbc_data_dep.R
\name{importBC_data_}
\alias{importBC_data_}
\title{Main data retrieval function}
\usage{
importBC_data_(
  parameter_or_station,
  years = NULL,
  use_openairformat = TRUE,
  use_ws_vector = FALSE,
  pad = TRUE
)
}
\arguments{
\item{years}{the years that will be retrieved. For sequence, use 2009:2015. For non
sequential years, use c(2010,2015,2018)
If not declared, the current year will be used}

\item{use_openairformat}{is boolean,if TRUE, output is compatible with openair. Apples only to station queries}

\item{use_ws_vector}{use vector wind speed? default is FALSE, if TRUE and use_openairformat is TRUE, ws is the vector wind speed}

\item{pad}{default is TRUE. if FALSE, it removes all NaNs}

\item{caaqs}{default is FALSE. If TRUE, it adds fields for TFEE, \n
and merges station names}
}
\description{
Deprecated since envair 0.2.1.000. use importBC_data() instead
This was original function since envair 0.1.0.000
It is still the back-end to retrieving data
}
