% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBC_data_avg.R
\name{importBC_data_avg0}
\alias{importBC_data_avg0}
\title{Back end function, see importBC_data_avg for main script}
\usage{
importBC_data_avg0(
  parameter,
  years = NULL,
  averaging_type = NULL,
  data_threshold = 0.75,
  flag_TFEE = FALSE,
  merge_Stations = FALSE
)
}
\description{
This function is capable of compoung averaging_type (e.g., "annual 98p d1hm")
But not ideal for multiple years or parameters
}
