% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBC_data_avg.R
\name{importBC_data_avg_}
\alias{importBC_data_avg_}
\title{Backend function of importBC_data_avg_}
\usage{
importBC_data_avg_(df, averaging_type, data_threshold = 0.75)
}
\description{
Note that averaging_type here is the simple (not compounded) format.
}
