% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventingBC_bulletin.R
\name{importECCC_forecast}
\alias{importECCC_forecast}
\title{GET ECCC FORECAST function}
\usage{
importECCC_forecast(parameter = NULL)
}
\arguments{
\item{parameter}{is a vector string of AQHI, PM25, NO2, O3, PM10. It will include all forecase if NULL}
}
\description{
This function retrieves latest forecast from ECCC models
}
