% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{isdf_DST}
\alias{isdf_DST}
\title{Retrieve the local time using PST input}
\usage{
isdf_DST(lst_DATEPST)
}
\arguments{
\item{lst_DATEPST}{is list or dataframe column containing the datetime in PST}
}
\description{
Retrieve the local time using PST input
}
