% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{isit_DST}
\alias{isit_DST}
\title{Determine if date is DST in BC or not}
\usage{
isit_DST(datetimePST = NULL, dateoutput = FALSE)
}
\arguments{
\item{datetimePST}{can be character string or datetime in PST
if NULL, is the current datetime.}

\item{dateoutput}{If FALSE returns a boolean. If TRUE, It outputs local time in string format i}
}
\description{
Determine if date is DST in BC or not
}
