% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBC_stations.R, R/listBC_stations_dep.R
\name{list_parameters}
\alias{list_parameters}
\title{List available parameters}
\usage{
list_parameters()

list_parameters()
}
\description{
This function lists the parameters that are available for retrieval

This function lists the parameters that are available for retrieval
}
\examples{
list_parameters()

list_parameters()

}
