% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBC_stations.R
\name{listBC_stations}
\alias{listBC_stations}
\title{List BC Air Quality Monitoring Stations}
\usage{
listBC_stations(year = NULL)
}
\arguments{
\item{year}{the year where station details are retrieved from. Defaults to current year if undefined
This is not a vector}
}
\description{
This function retrieves latest station details or deteails during specific year from the ftp feed
}
\examples{
listBC_stations()
listBC_stations(2015)

}
