% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listBC_stations.R, R/listBC_stations_dep.R
\name{listBC_stations_}
\alias{listBC_stations_}
\title{Main listBC_stations() function}
\usage{
listBC_stations_(year = NULL)

listBC_stations_(year = NULL)
}
\description{
this was created to ensure output is up-to-date
when user specifies a year

this was created to ensure output is up-to-date
when user specifies a year
}
