% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caaqs_stn_history.R
\name{merge_STATIONS}
\alias{merge_STATIONS}
\title{Merges station names for reporting purposes}
\usage{
merge_STATIONS(
  df,
  station_column = NULL,
  data_column = NULL,
  instrument_column = NULL,
  date_column = NULL,
  param_column = NULL,
  is_time_ending = TRUE
)
}
\arguments{
\item{df}{is the dataframe containing the data}

\item{station_column}{is a string defining the station name.
By default it will be the column containg "station" or "site" in the name}

\item{data_column}{is the column containing the pollutant data}

\item{instrument_column}{is the column defining the instrument type}

\item{date_column}{is the column containing the date.
Note that if this is date/time column, is_time_ending should be specified}

\item{param_column}{is the column defining the pollutant or parameter}

\item{is_time_ending}{is either TRUE/FALSE. default is TRUE.
time_ending means that the start of the day is AFTER and NOT including midnight.}
}
\description{
This function combines station names for CAQQS-reporting purposes
It uses the list created in the FTP server 'ftp://ftp.env.gov.bc.ca/pub/outgoing/AIR/CAAQS/BC_CAAQS_station_history.xlsx'
}
\examples{
df <- importBC_data('pm25',2019)
df <- merge_STATIONS(df)
}
