% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paddatafunction.R
\name{pad_data}
\alias{pad_data}
\title{Pad data to completely represent entire years}
\usage{
pad_data(
  df,
  date_time = NULL,
  values = NULL,
  time_ending = TRUE,
  add_DATETIME = NULL
)
}
\arguments{
\item{df}{is the dataframe containing date_time parameters, and values columns}

\item{date_time}{defines the date_time column. Default is DATE_PST}

\item{values}{defines the columns with the reported measurements. If NULL, it wil use RAW_VALUE and ROUNDED_VALUE}

\item{time_ending}{if TRUE means date starts AFTER and EXCLUDES midnight}

\item{add_DATETIME}{if TRUE, means to force and include DATETIME}
}
\description{
This function inserts fillers to data set. The function can identify separate columns of DATE and TIME
}
