% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbc_data.R
\name{process_data_history}
\alias{process_data_history}
\title{CHeck the history of station and instruments
and updated the data}
\usage{
process_data_history(df_data)
}
\description{
CHeck the history of station and instruments
and updated the data
}
