% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_process.R
\name{process_ftp_files}
\alias{process_ftp_files}
\title{Download a list of FTP files}
\usage{
process_ftp_files(url_list)
}
\arguments{
\item{url_list}{is a vector containing list of URLs to download}
}
\description{
uses parallel CPU processing to perform simultaneous downloads
}
