% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventingBC_bulletin.R
\name{read_venting_file}
\alias{read_venting_file}
\title{Read venting file, extract information, tabulate}
\usage{
read_venting_file(url)
}
\arguments{
\item{url}{is the URL for the text bulletin}
}
\description{
Read venting file, extract information, tabulate
}
