% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{RENAME_COLUMN}
\alias{RENAME_COLUMN}
\title{RENAME_COLUMN function}
\usage{
RENAME_COLUMN(data.station, colname.orig, colname.new = NULL)
}
\arguments{
\item{data.station}{dataframe input}

\item{colname.orig}{string vector containing the column to be deleted or renamed}

\item{colname.new}{string vector containing the new name for the column. if NULL, it deletes the specified column}
}
\description{
This function renames or deletes the column of a dataframe based on its name
}
\keyword{RENAME_COLUMN()}
\keyword{dataframe}
\keyword{rename}
