% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{RENAME_COLUMN_untidy}
\alias{RENAME_COLUMN_untidy}
\title{RENAME_COLUMN function}
\usage{
RENAME_COLUMN_untidy(
  data.station,
  colname.orig,
  colname.new = NULL,
  quiet = TRUE
)
}
\arguments{
\item{data.station}{dataframe input}

\item{colname.orig}{string vector containing the column to be deleted or renamed}

\item{colname.new}{string vector containing the new name for the column. if NULL, it deletes the specified column}
}
\description{
This function renames or deletes the column of a dataframe based on its name
this was created to prevent any error when the column name does not exist
It is, at the same time, able to delete column
}
\keyword{RENAME_COLUMN()}
\keyword{dataframe}
\keyword{rename}
