% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairaqhi.R, R/envairfunctions.R,
%   R/publish_aqhi.R
\name{round2}
\alias{round2}
\title{rounding off function
#
This function makes a simple rounding off}
\usage{
round2(x, n = 0)

round2(x, n = 0)

round2(x, n = 0)
}
\arguments{
\item{x}{the as.numeric input}

\item{n}{the number of decimal places for resulting output
round2()}

\item{x}{the as.numeric input}

\item{n}{the number of decimal places for resulting output
round2()}

\item{x}{the as.numeric input}

\item{n}{the number of decimal places for resulting output
round2()}
}
\description{
rounding off function
#
This function makes a simple rounding off

rounding off function
#
This function makes a simple rounding off

rounding off function
#
This function makes a simple rounding off
}
