% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{roundall}
\alias{roundall}
\title{round off columns to the precision}
\usage{
roundall(data, cols_round, n = 0)
}
\arguments{
\item{data}{is a data frame}

\item{cols_round}{is vector of string listing the columns to round to specified precision}

\item{n}{is the precision or number of decimal places}
}
\description{
This function rounds off specified columns in a dataframe to a specified precision
}
\examples{
df <- importBC_data_avg('pm25',2020,averaging_type = 'annual mean 24h')
df_rounded <- roundall(df_rounded,cols_round = c('rounded_mean_24h'),n=1)
}
