% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_station_data.R, R/email_aqhi.R,
%   R/envairaqhi.R, R/envairfunctions.R, R/ftp_reexport.R
\name{RUN_PACKAGE}
\alias{RUN_PACKAGE}
\title{Install and Load specific libraries that are useful for the envair package}
\usage{
RUN_PACKAGE(packages = c("dplyr", "ggplot2", "reshape", "lazyeval",
  "zoo", "DataCombine", "data.table", "fasttime", "readr", "RCurl",
  "tidyr", "lubridate"), lib.path = NULL)

RUN_PACKAGE(packages = c("dplyr", "ggplot2", "reshape", "lazyeval",
  "zoo", "DataCombine", "data.table", "fasttime", "readr", "RCurl",
  "tidyr", "lubridate"), lib.path = NULL)

RUN_PACKAGE(packages = c("dplyr", "ggplot2", "reshape", "lazyeval",
  "zoo", "DataCombine", "data.table", "fasttime", "readr", "RCurl",
  "tidyr", "lubridate"), lib.path = NULL)

RUN_PACKAGE(packages = c("dplyr", "ggplot2", "reshape", "lazyeval",
  "zoo", "DataCombine", "data.table", "fasttime", "readr", "RCurl",
  "tidyr", "lubridate"), lib.path = NULL)

RUN_PACKAGE(packages = c("dplyr", "ggplot2", "reshape", "lazyeval",
  "zoo", "DataCombine", "data.table", "fasttime", "readr", "RCurl",
  "tidyr", "lubridate"), lib.path = NULL)
}
\arguments{
\item{packages}{vector string listing the packages}

\item{lib.pack}{path of library, defaults to current lib paths
RUN_PACKAGE()}

\item{packages}{vector string listing the packages}

\item{lib.pack}{path of library, defaults to current lib paths
RUN_PACKAGE()}

\item{packages}{vector string listing the packages}

\item{lib.pack}{path of library, defaults to current lib paths
RUN_PACKAGE()}

\item{packages}{vector string listing the packages}

\item{lib.pack}{path of library, defaults to current lib paths
RUN_PACKAGE()}
}
\description{
RUN_PACKAGE()

This function isntalls and loads packages if needed

This function isntalls and loads packages if needed

This function isntalls and loads packages if needed

This function isntalls and loads packages if needed
}
