% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{SAVE_TO_FILE}
\alias{SAVE_TO_FILE}
\title{Save dataframe to a file in a specified location}
\usage{
SAVE_TO_FILE(
  data_to_save,
  filename = NULL,
  path.target,
  path.temp = NULL,
  file_only = FALSE
)
}
\arguments{
\item{data_to_save}{is the dataframe to save into a file, or
for file_only, it is the file to transfer}

\item{filename}{is string defining the name of the file,
if NULL, it keeps the namefrom data_to_save, works only if file_only = TRUE}

\item{path.target}{is string defining the path where the file will be saved}

\item{path.temp}{default NULL, is where the file is temporarily saved. if null, file is saved in working directory}

\item{file_only}{is true, it performs a file transfer to FTP}
}
\description{
This function saves the dataframe into a file
}
