% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caaqs_stn_history.R, R/importbc_data.R
\name{space_to_underscore}
\alias{space_to_underscore}
\title{replace space in string to underscore or space
elso ensure removal of non-visible characters}
\usage{
space_to_underscore(x, replace = "_", collapse = TRUE)

space_to_underscore(x, replace = "_", collapse = TRUE)
}
\arguments{
\item{x}{is the string}

\item{replace}{is the character that the space will be replaced to}

\item{collapse}{removes double space}
}
\description{
replace space in string to underscore or space
elso ensure removal of non-visible characters

replace space in string to underscore or space
elso ensure removal of non-visible characters
}
