% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventingBC_bulletin.R
\name{ventingBC_bulletin}
\alias{ventingBC_bulletin}
\title{generate the venting index bulletin for BC env}
\usage{
ventingBC_bulletin(date.start = NULL, savefile = NULL, output = "html")
}
\arguments{
\item{date.start}{is the date of the venting data. if NULL, it retrieves the latest available}

\item{savefile}{is the location where the file will be saved. If null it will default to locations
defined in default_html, and default_text locations based on output
If an ftp link is defined, envair.key with write access details to FTP server is needed
If savefile is terminated in '/', another folder for the data year (~./2019/EC_) will be generated and
the}

\item{output}{is either 'html' or 'ftp'. If 'html', it will create html file into envistaweb
if 'ftp' then it will save into the Open Data Portal FTP page
if 'csv' it creates the csv file that populates the http://www.env.gov.bc.ca/epd/bcairquality/aqo/csv/VentingIndex.csv}
}
\description{
This function retrieves venting data from ECCC datamart
to generate the text bulletin
function here was developed to run in the DAS server (PARTICULATE.dmz)
which saves the resulting venting.html file into E:/WebSites/wwwroot/Web2016/aqo/files/bulletin/
and will show up in
}
\examples{
ventingBC_bulletin()

}
