% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_station_data.R, R/email_aqhi.R,
%   R/envairaqhi.R, R/envairfunctions.R
\name{COLUMN_REORDER}
\alias{COLUMN_REORDER}
\title{REORDER COLUMNS}
\usage{
COLUMN_REORDER(data.input, columns = c(""))

COLUMN_REORDER(data.input, columns = c(""))

COLUMN_REORDER(data.input, columns = c(""))

COLUMN_REORDER(data.input, columns = c(""))
}
\arguments{
\item{data.input}{the input data frame}

\item{columns}{the vector of strings listing the column names in desired order. Columns not listed are added at the end
COLUMN_REORDER()}

\item{data.input}{the input data frame}

\item{columns}{the vector of strings listing the column names in desired order. Columns not listed are added at the end
COLUMN_REORDER()}

\item{data.input}{the input data frame}

\item{columns}{the vector of strings listing the column names in desired order. Columns not listed are added at the end
COLUMN_REORDER()}

\item{data.input}{the input data frame}

\item{columns}{the vector of strings listing the column names in desired order. Columns not listed are added at the end
COLUMN_REORDER()}
}
\description{
This function reorders the columns of a dataframe based on the order it is specified in columns

This function reorders the columns of a dataframe based on the order it is specified in columns

This function reorders the columns of a dataframe based on the order it is specified in columns

This function reorders the columns of a dataframe based on the order it is specified in columns
}
