% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_process.R
\name{download_file}
\alias{download_file}
\title{Downloads files into the URL}
\usage{
download_file(file_url)
}
\arguments{
\item{file_url}{defines the URLs and destinaton}
}
\description{
Downloads files into the URL
}
