% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importbc_data.R
\name{download_files}
\alias{download_files}
\title{Download a list of FTP files}
\usage{
download_files(url_list)
}
\arguments{
\item{url_list}{is a vector containing list of URLs to download}
}
\description{
uses parallel CPU processing to perform simultaneous downloads
}
