% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_captures.R
\name{get_captures}
\alias{get_captures}
\title{Calculate the data capture for the station and parameter}
\usage{
get_captures(param, years = NULL, adjust = FALSE, stop_at_present = TRUE)
}
\arguments{
\item{param}{is the pollutant(param). Use list_parameters() to list availble parameters.
You can use vector for multi-pollutant query.
and you can also use the data itself. Data must be retrieved form an importBC_data()}

\item{years}{is the year or years. You can use vectors to query multiple years. If NULL, it will be the current year}

\item{adjust}{default is FALSE. if TRUE (not working), it will combine some stations/instruments
based on the merging requirement applies in CAAQS}

\item{stop_at_present}{default is TRUE. It will remove values after the present date/time
The present date/time is determined from the latest PM2.5 data}
}
\description{
NOTE: This will output the valid days, valid days in quarter, valid
}
\examples{
get_captures('o3',years = 2017:2020)
df <- importBC_data('o3',years = 2017:2020)
get_captures(df)

}
