% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_caaqs_stn_history.R
\name{get_station_history}
\alias{get_station_history}
\title{Retrieve station history for CAAQS-reporting}
\usage{
get_station_history(ExcelURL = NULL)
}
\arguments{
\item{ExcelURL}{is the complete URL of the excel file containing station history data
leave NULL to use known location of FTP at 'ftp://ftp.env.gov.bc.ca/pub/outgoing/AIR/CAAQS/BC_CAAQS_station_history.xlsx'}
}
\description{
This function retrieves station and instrument history for CAAQS-reporting purposes
The output is a dataframe showing the station and instrument start and end dates, and merged name
}
