% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Calculate the statistucs for the pollutants based on CAAQS}
\usage{
get_stats(param, years = NULL)
}
\arguments{
\item{param}{is the}

\item{datetime}{is a string defining the datetime field. This field shoule be in time-ending format.}
}
\description{
Note that these are based on the metrics defined by the CCME
Guidance Document of Achievement Determination
}
