% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventingBC_bulletin.R
\name{GET_VENTING_ECCC}
\alias{GET_VENTING_ECCC}
\title{Retrieves the venting index FLCN 39 from ECCC data mart}
\usage{
GET_VENTING_ECCC(dates = NULL)
}
\arguments{
\item{dates}{datetime or string. can be a vector
if left undefined, it uses the current date}
}
\description{
This function connects to ECCC datamart and retrieves venting data
}
\examples{
GET_VENTING_ECCC()
GET_VENTING_ECCC(dates = seq(from = lubridate::ymd('2021-01-01'),
       to = lubridate::ymd('2021-05-01'), by = 'day'))

}
