% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_station_data.R, R/importbc_data.R
\name{importBC_data}
\alias{importBC_data}
\title{Function copied from importbc_data.R
copied 2019-10-16 version
Function for retrieving ENV air quality data from all stations and specific year}
\usage{
importBC_data(parameter_or_station, years = NULL,
  use_openairformat = TRUE)
}
\arguments{
\item{parameter_or_station}{vector list of air quality parameters or station (automatic).
use the function listBC_stations() to get a detailed list of BC Air Quality Monitoring stations
if there is not sation match, it will use the search item as station keyword
so that user can just enter 'Prince George' and it will search for all stations
with Prince  George in the station name (Prince George Plaza 400, Prince George Glenview)

List of parameters can be obtained using list_parameters() command
List of stations can be retrieved using listBC_stations() command}

\item{years}{the years that will be retrieved. For sequence, use 2009:2015. For non
sequential years, use c(2010,2015,2018)
If not declared, the current year will be used}

\item{use_openairformat}{is boolean,if TRUE, output is compatible with openair}

\item{parameters}{the air quality parameters. can be PM25, PM10, SO2, TRS, H2S, etc. This is not case sensitive.
to specify more that one parameter, use c('SO2','co')
List of parameters can be obtained from the FTP link where data is extracted from:
'ftp://ftp.env.gov.bc.ca/pub/outgoing/AIR/AnnualSummary/'}

\item{year}{the years that will be retrieved. For sequence, use 2009:2015. For non
sequential years, use c(2010,2015,2018)
If not declared, the current year will be used

importbc_paramdata()}
}
\description{
This function retrieves parameter data from the FTP open data portal
Data includes verified and unverified data

This function retrieves station or parameter hourly data from the BC open data portal
Data includes verified and unverified data
}
\examples{
importBC_data('Prince George Plaza 400')
importBC_data('pm25',2015:2016,use_openairformat = FALSE)

}
