% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBC_data_avg.R
\name{importBC_data_avg}
\alias{importBC_data_avg}
\title{Average the imported data function}
\usage{
importBC_data_avg(
  parameter,
  years = NULL,
  averaging_type = NULL,
  data_threshold = 0.75,
  flag_tfee = FALSE,
  merge_Stations = FALSE
)
}
\arguments{
\item{parameter}{is the parameter to average.use list_parameters() for list of parameters.
parameter can also be a dataframe from  importBC_data(). If dataframe, it will process and retrieve average.}

\item{years}{specifies the years to include. Use vector for multiple years. If NULL, it uses the current year.}

\item{averaging_type}{is the sub-annual values of either "1-hour", "24-hour", "8-hour","d1hm", "d8hm"
You can also request annual summaries, percentiles. Add "annual <annual stat> <on what data>": 'annual 98p 24h','annual 99p d1hm','annual xxp 1hr';'annual mean 1-hr';
'annual mean 24-hr';'annual 4th d1hm'
It can also calculate the number of exceedance to a specified value "exceedance 25.00 d1hm".
Function will round off based on the precision of the number entered after exceedance
If NULL, it applies either the 1-hour, 8-hour, or 24-hour, depending on typical default for}

\item{data_threshold}{value between 0 to 1. This refers to the data capture requirements.
Data will less than that data_threshold are excluded from the output.
If data_threshold=0, it will include ALL values of that averaging type and will include}

\item{flag_tfee}{default is FALSE. If TRUE, it will evaluate data with TFEE adjustment and include it in the result.
The resultig column will have _TFEE for the TFEE-adjusted result}

\item{merge_Stations}{default is FALSE. If TRUE, data from stations and their alternatives are merged. Alternative stations
are defined when station has relocated to a nearby area}
}
\value{
A wide dataframe listting the date, station name, instrument, parameter, and then the raw and rounded values with underscore
that specified if 24hour, and if adjusted for tfee
}
\description{
This function averages the imported data
}
\examples{
importBC_data_avg('o3')   #retrieves the daily 8-hour maximum for current year
importBC_data_avg('o3',years = 2015, averaging_type = 'd1hm',data_threshold = 0)   #displays the daily 1-hour maximum for ozone in 2015. It will display all data even if less than 75\% of the day are not available.

}
