% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBC_data_avg.R
\name{importBC_data_avg_}
\alias{importBC_data_avg_}
\title{Backend function of importBC_data_avg_
These will calculate the daily 1-hour maximum values
Note that averaging_type here is the simple (not compounded) format}
\usage{
importBC_data_avg_(
  parameter,
  years = NULL,
  averaging_type = NULL,
  data_threshold = 0.75
)
}
\description{
Backend function of importBC_data_avg_
These will calculate the daily 1-hour maximum values
Note that averaging_type here is the simple (not compounded) format
}
