% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{isit_DST}
\alias{isit_DST}
\title{Determine if date is DST in BC or not}
\usage{
isit_DST(datetime = Sys.time())
}
\arguments{
\item{datetime}{is the date and time}
}
\description{
Determine if date is DST in BC or not
}
