% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envairfunctions.R
\name{round2}
\alias{round2}
\title{rounding off function}
\usage{
round2(x, n = 0)
}
\arguments{
\item{x}{the as.numeric input}

\item{n}{the number of decimal places for resulting output
round2()}
}
\description{
This function makes a simple rounding off based on ISO17025
}
